efi: Add tee-based EFI variable driver
When the underlying storage is owned by the secure world, accessing EFI
variables is straightforward and done via EFI Runtime Variable Services.
In this case, critical variables for system integrity and security are
normally stored in the dedicated secure storage and only accessible from
the secure world.

On the other hand, most small embedded devices don't have the special
dedicated secure storage. The eMMC device with an RPMB partition is
becoming more common, we can use an RPMB partition to store the EFI
Variables.

The eMMC device is typically owned by the non-secure world (Linux in
this case). There is an existing solution utilizing eMMC RPMB partition
for EFI Variables, it is implemented by interacting with TEE (OP-TEE in
this case), StandaloneMM (as EFI Variable Service Pseudo-TA *), eMMC
driver and tee-supplicant in user space. The last piece is the tee-based
variable access driver to interact with TEE and StandaloneMM.

So let's add the kernel functions needed.

This feature is implemented as a kernel module.  StMM pseudo-TA has
TA_FLAG_DEVICE_ENUM_SUPP flag when registered to OP-TEE so that this
tee_stmm_efi module is probed after tee-supplicant starts, since
"SetVariable" EFI Runtime Variable Service requires to interact with
tee-supplicant.

[ardb: the main reason this needs to live in the kernel rather than in
       the non-secure EFI runtime services code exposed by the firmware
       is that RPMB access via OP-TEE is asynchronous, and so it relies
       on the OP-TEE runtime in the kernel for the scheduling and the
       completion, and firmware does not have access to that]

* TA == Trusted Application

Co-developed-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>
Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>
Signed-off-by: Masahisa Kojima <masahisa.kojima@linaro.org>
[ardb: drop unused tee_query_variable_info()]
Signed-off-by: Ard Biesheuvel <ardb@kernel.org>