maple_tree: reduce resets during store setup
mas_prealloc() may walk partially down the tree before finding that a
split or spanning store is needed.  When the write occurs, relax the logic
on resetting the walk so that partial walks will not restart, but walks
that have gone too far (a store that affects beyond the current node)
should be restarted.

Link: https://lkml.kernel.org/r/20230612203953.2093911-16-Liam.Howlett@oracle.com
Signed-off-by: Liam R. Howlett <Liam.Howlett@oracle.com>
Cc: Peng Zhang <zhangpeng.00@bytedance.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>