dt-bindings: display/msm: sc7180-dpu: Describe SM6125
SM6125 is identical to SM6375 except that while downstream also defines
a throttle clock, its presence results in timeouts whereas SM6375
requires it to not observe any timeouts.  This is represented by
reducing the clock array length to 6 so that it cannot be passed.  Note
that any SoC other than SM6375 (currently SC7180 and SM6350) are
unconstrained and could either pass or leave out this "throttle" clock.

Signed-off-by: Marijn Suijten <marijn.suijten@somainline.org>
Reviewed-by: Rob Herring <robh@kernel.org>
Patchwork: https://patchwork.freedesktop.org/patch/544521/
Link: https://lore.kernel.org/r/20230627-sm6125-dpu-v2-6-03e430a2078c@somainline.org
Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>