NFSD: Enforce flush-on-close for write delegations
We've found that when the Linux client unlinks a file for which it
holds a write delegation, it silly-renames it first then unlinks it.
This causes problems for applications that concurrently read the
parent directory: the application sees .nfs files and that prevents
the directory's immediately removal.

As a workaround, force clients to flush-on-close when unlinking a
file that has a write delegation. This has performance implications,
of course, so it is not an ultimate solution.

Suggested-by: Dai Ngo <dai.ngo@oracle.com>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>