firmware: arm_ffa: Add interface to send a notification to a given partition
The framework provides an interface to the sender endpoint to specify
the notification to signal to the receiver endpoint. A sender signals
a notification by requesting its partition manager to set the
corresponding bit in the notifications bitmap of the receiver.

Expose the ability to send a notification to another partition.

Link: https://lore.kernel.org/r/20230929-ffa_v1-1_notif-v3-11-c8e4f15190c8@arm.com
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>