btrfs: introduce the temp-fsid feature
Btrfs doesn't currently support to mount 2 different devices holding the
same filesystem - the fsid is exposed as a unique identifier by the
driver. This case is supported though in some other common filesystems,
like ext4.

Supporting the same-fsid mounts has the advantage of allowing btrfs to
be used in A/B partitioned devices, like mobile phones or the Steam Deck
for example. Without this support, it's not safe for users to keep the
same "image version" in both A and B partitions, a setup that is quite
common for development, for example. Also, as a big bonus, it allows fs
integrity check based on block devices for RO devices (whereas currently
it is required that both have different fsid, breaking the block device
hash comparison).

Such same-fsid mounting is hereby added through the usage of the
filesystem feature "temp-fsid" - when such feature is used, btrfs
generates a random fsid for the filesystem and leverages the long-term
present metadata_uuid infrastructure to enable the usage of this
secondary "virtual" fsid, effectively requiring few non-invasive
changes to the code and no new potential corner cases.

In order to prevent more code complexity and corner cases, the
temp-fsid feature is not allowed when the metadata_uuid flag is
present on the fs, or if the device is on fsid-change state. Device
removal/replace is also disabled for filesystems presenting the feature.

Cc: Anand Jain <anand.jain@oracle.com>
Suggested-by: John Schoenick <johns@valvesoftware.com>
Suggested-by: Qu Wenruo <wqu@suse.com>
Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>
Signed-off-by: David Sterba <dsterba@suse.com>