KVM: arm64: FFA: Remove access of endpoint memory access descriptor array
FF-A v1.1 removes the fixed location of endpoint memory access descriptor
array within the memory transaction descriptor structure. In preparation
to remove the ep_mem_access member from the ffa_mem_region structure,
provide the accessor to fetch the offset and use the same in FF-A proxy
implementation.

The accessor take the boolean argument that indicates if the memory access
descriptor versions is v1(old format) or not. Currently it is set true as
FF-A proxy supports only v1.0

Cc: Marc Zyngier <maz@kernel.org>
Cc: Oliver Upton <oliver.upton@linux.dev>
Cc: Will Deacon <will@kernel.org>
Cc: Quentin Perret <qperret@google.com>
Link: https://lore.kernel.org/r/20230929-ffa_v1-1_notif-v3-14-c8e4f15190c8@arm.com
Signed-off-by: Sudeep Holla <sudeep.holla@arm.com>