gfs2: More aggressive page fault-in
To prevent deadlocks, gfs2 is performing reads and writes with page
faults disabled.  When the buffer to read from or write to is not
resident in memory, the read or write will return an error, the buffer
will be faulted in, and the the read or write will be retried.

The function at the core of those retries is should_fault_in_pages(); it
computes how much memory to fault in at a time.  So far, this function
limited that size to 1 MiB and tried to scale the window size back based
on memory pressure using a crude heuristic.  This could cause a lot more
requests to be issued at the block layer.  In addition, when
asynchronous direct reads or writes are split, all but the last chunk
will actually be completed synchronously.  Depending on the application
and the storage used, this can be quite costly.

To avoid these problems, always try to fault in the entire buffer at
once.  This will result in more memory pressure when applications are
carrying out very large reads or writes, but that can be optimized at
the application level when necessary.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>