selftests/vDSO: don't #include sodium header in chacha test
With changes in various kernel headers, there's a clash if we include
the libsodium header. This is especially hard to fix with build flags,
because we're compiling the assembly together with the C file, for
testing. Instead, just add the single library prototype we're using in
the file and don't bother including the real header.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>