dt-bindings: arm64: mediatek: Add kukui-jacuzzi-cerise board
Cerise is known as ASUS Chromebook CZ1.
Stern is known as ASUS Chromebook Flip CZ1.

They are almost identical. The only difference is that Cerise is a
clamshell device without touchscreen and Stern is a convertible device.

Although the commercial names are different, the bootloader still
expects these cerise compatible strings even on stern.

Signed-off-by: Hsin-Te Yuan <yuanhsinte@chromium.org>
Acked-by: Krzysztof Kozlowski <krzysztof.kozlowski@linaro.org>
Link: https://lore.kernel.org/r/20240731-jacuzzi_dt-v3-2-1c4314e8962f@chromium.org
Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>